requirements.txt
# Core Django
Django>=5.0.1
Pillow>=10.1.0
python-decouple>=3.8
pytz>=2023.3

# Database - use psycopg2-binary for Railway
psycopg2-binary>=2.9.9

# Caching & Task Queue
redis>=5.0.1
django-redis>=5.4.0
celery>=5.3.4
django-celery-beat>=2.5.0

# Payment Gateway
razorpay>=1.4.1

# Email & SMS Notifications
twilio>=8.10.0

# Google Calendar API (PRO plan only)
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0

# AI Features (OpenAI)
openai>=1.3.5

# Production Server
gunicorn
whitenoise

# Production Email Services (Optional)
sendgrid>=6.11.0
django-ses>=3.5.0

# AWS S3 for Media Files (Optional)
boto3>=1.34.0
django-storages>=1.14.2

# Monitoring & Error Tracking (Optional)
sentry-sdk>=1.39.1

# Security & Performance
django-cors-headers>=4.3.1
django-ratelimit>=4.1.0

# Database File Storage
django-db-file-storage>=0.5.6

# Forms
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.1

# Database Backup (Optional)
django-dbbackup>=4.0.2

# Environment Management
python-dotenv>=1.0.0

# Math template filters
django-mathfilters>=1.0.0

# DNS lookups (required for domain verification)
# DNS lookups (required for domain verification)
dnspython>=2.4.0
dnspython>=2.4.0
